refactor: player 패키지 추가
